+++ packages/isaacscript-common/src/types/Range.ts
@@ -1,13 +1,15 @@
+import { NaturalNumbersEqualToOrLessThan } from "./NaturalNumbersEqualToOrLessThan";
 import { NaturalNumbersLessThan } from "./NaturalNumbersLessThan";
 
 /**
- * Helper type to get a range of integers between low and high. It is inclusive on the lower end and
- * exclusive on the higher end.
+ * Helper type to get a range of integers. It is inclusive on both ends.
+ *
+ * For example, `Range<3, 5>` will return `3 | 4 | 5`.
  *
  * From:
  * https://stackoverflow.com/questions/39494689/is-it-possible-to-restrict-number-to-a-certain-range
  */
 export type Range<Low extends number, High extends number> = Exclude<
-  NaturalNumbersLessThan<High>,
+  NaturalNumbersEqualToOrLessThan<High>,
   NaturalNumbersLessThan<Low>
 >;