+++ packages/isaacscript-common/src/types/TupleWithMaxLength.ts
@@ -1,0 +1,13 @@
+import { NaturalNumbersEqualToOrLessThan } from "./NaturalNumbersEqualToOrLessThan";
+
+/**
+ * Helper type that validates that a tuple does not have a length greater than N.
+ *
+ * For example, `TupleWithMaxLength<string, 3>` will allow string tuples of size 0, 1, 2, or 3.
+ */
+export type TupleWithMaxLength<T, MaxLength extends number> = (
+  | T[]
+  | readonly T[]
+) & {
+  length: NaturalNumbersEqualToOrLessThan<MaxLength>;
+};